Basic PingPong
This is a basic implementation of the Atari Pong game, in which a ball is balanced back and forth on two graphically designed slabs. It is a 2-player game in which the ball should be balanced such that, it does not touch the court side. A complete track of scores is present along with 3 classes.